core: increment vm/templates db generation on pool ver upgrade
When upgrading the storage pool comptability version to one that
supports the OvfStoreOnAnyDomain feature, the vms/templates db
generation should be incremented so that ovf configuration will be
generated and saved for those vms/templates.

Without it being done on the next run of the ovf update process
for a domain we might store ovf with missing ovf configurations
(as they were never generated and saved).

Change-Id: Ief1a984a7d275d558bbe74bbca6bd49326267e43
Signed-off-by: Liron Aravot <laravot@redhat.com>