findbugs: Make inner classes static where possible
This patch makes all inner classes static, where possible[1]. The
main advantage of making an inner class static is the load removed from
the classloader. This way, an inner class is created per enclosing CLASS
instead of per enclosing INSTANCE.

This patch also enables a FindBugs check to ensure that such mistakes
are not introduced after this patch.

[1] A notable exception to this rule are inner classes in classes that
are GWT-compiled. Any inner class with the @GenEvent annotation is
converted to a top-level class with the same modifiers by GWT, and
therefore can't be declared static.

Change-Id: Ifbceb1baad596211580c1289f49f7d12e33a8ecb
Signed-off-by: Allon Mureinik <amureini@redhat.com>