webadmin: Aggregate error messages in Frontend.runMultipleActions()
Added the ability to aggregate error messages when invoking
Frontend.runMultipleActions(). This is useful for cases where batch
operations need to be run serially, and even though one failure
shouldn't prevent the subsequent operations from executing, errors
still need to be aggregated to be shown in one error dialog.

To process the multiple error messages of the failed operations, the
IFrontendEventsHandler.runMultipleActionFailed() method was extended
slightly to be able to handle failures that aren't canDoAction()
failures.

Change-Id: I0236b7d9316dff3d9bd22f2127318eefc37bbda1
Bug-Url: https://bugzilla.redhat.com/1048752
Signed-off-by: Lior Vernia <lvernia@redhat.com>