engine: Adding a down event for HA Reservation
Adding a down event to HA Reservation, this event will be triggered
only in case the current status of the Cluster was changed from not HA safe to HA safe.
This is done to ensure that the user is notified that his action
(freeing resources) have made the Cluster HA safe.

Change-Id: Iebc82d3cc190d439b77e45e7e6b3eb7451f71c3c
Bug-Url: https://bugzilla.redhat.com/1076211
Signed-off-by: Kobi Ianko <kobi@redhat.com>