aaa: Changing API to 3.3 behavior
Changing back the behavior of the API to 3.3 behavior.
Changing also the behavior of internal id allocation -
this is cancelled, and internal id (primary key) will always be the guid
returned from the directory.
Although in 3.5 we will have directories that are non ldap,
hence can have ids that are not GUID, we would like to handle
the user id representation differently, and this is a first stage.

Topic: AAA
Change-Id: Iff51fb9df33b4b679eba5b1c3ef2bea6ea7b3e07
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>