core: last ovf store update
In a case in which the last ovf store was previously updated, we should
attempt to update it only if update to any of the previous domains
was succesfull.
Otherwise, we can attempt to update it immediately as it doesn't contain
any ovf data so there's no point to consider the other updates results.

Change-Id: I21491330681c333faf9e97b3d348d2af8c2b2fb5
Signed-off-by: Liron Aravot <laravot@redhat.com>