core: Fix AuditLog check for existing log entries
The audit log check for existing element in the cache,
If the entry exist, then there is no need for another log message.
This is how the anti flooding mechanism works.
For that we use Infinispan cache and the putIfAbsent method,
which updates the lifespan(time to keep in cache).
For our use of the cache, we don't want this update.

The fix changes the putIfAbsent and replace it with two othe API methods,
Contains and Put.

Change-Id: I94b777911eb1ff667be3dc8362bde5746d84824c
Bug-Url: https://bugzilla.redhat.com/1079920
Signed-off-by: Kobi Ianko <kobi@redhat.com>