engine: Avoid storing external network label in oVirt-engine DB
Network labels for oVirt-engine and for Neutron should not be mixed when
creating a network and labelling it. For oVirt-engine network only
oVirt-engine labels should be presented for newly created network
and for Neutron network only Neutron labels.
External network label should not be stored in oVirt-engine DB.
As soon as we won't store the external network labels it would never be
mixed up with the oVirt-engine managed network labels.

Change-Id: Ic3c041d6900a8af852f1d186c626be315d205be6
Bug-Url: https://bugzilla.redhat.com/1077132
Signed-off-by: Yevgeny Zaspitsky <yzaspits@redhat.com>