engine: event should contain the iface which exceeded defined threshold
A general event was reported per host when any of the host's NICs exceeded
the threshold (based TX or RX values). The event didn't contain the nics
name. And containd just the max TX or RX from all the interfaces.
This patch extends it to a specific event per interface.

Change-Id: I22b3510398ac6fa3acf93fcb596dd6ea10b26366
Bug-Url: https://bugzilla.redhat.com/1070667
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>