frontend: Correct RunOnce Iso selection
RunOnceDialog ignored persistent configuration of selected CD and
also didn't support unchecking the CD. (Frontend would display as not
attached CD but would run with attached CD anyway)

Now we support that by defaulting the RunVmParams#diskPath to be null
when the user explicitly uncheck the "Attach CD" checkbox.
We also pre-select the value of VmBase#isoPath in the iso listbox.

Change-Id: I714d8fad60673608d58ade4463c4abbcb5eb3f08
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1026306
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1054743
Signed-off-by: Martin Betak <mbetak@redhat.com>