engine: Send ifaceName and tpgt in connectStorageServer
Engine mistakenly sent initiatorName instead of ifaceName which overwrote
existing initiatorName configured in the host. Wrong initiatorName caused
host fail to perform iSCSI connection.

In case of tpgt the problem is terminology. Engine sent it as a "portal"
and its value is integer,  but from the iSCSI perspective portal is an
access point to one or more targets, while this integer actually
represents a tpgt - target portal group tag which is numeric representation
of which IP addresses a target is allowed to bind on.

Bug-Url: https://bugzilla.redhat.com/1078137
Change-Id: Ife467feb8aad9711da5d6a5df305e50fe6bf51e9
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>