core: rerun HA VMs on automatic fence using AutoStartVmsRunner
Before the engine sends fence request to VDSM, it tries to rerun the
highly available VMs that ran on the host by invoking RunVmCommand
directly. The problem is that we won't try to rerun the VM again if the
RunVmCommand fails (and in case the host which is being fenced in the
only one that was active in the cluster the command will fail for sure).
The solution is to rerun the highly available VMs using
AutoStartVmsRunner job in this flow as well.

Change-Id: Ic259ff22c4b6f24b09eb973de6e31ebc51aa966b
Bug-Url: https://bugzilla.redhat.com/1077706
Signed-off-by: Arik Hadas <ahadas@redhat.com>