core: set vm exit status to error on fencing
As part of fence operation, we set the VM status to down using
SetVmStatusVDSCommand. This command did not filled the exit status of
the VM which should be ERROR in this case. So this patch adds that
missing part by passing the exit status that should be set for the VM
while changing its status through the parameters of the command.

Change-Id: If541c6d33a7cfa310ba776e4bc7fa0f1e45aa077
Bug-Url: https://bugzilla.redhat.com/1077706
Signed-off-by: Arik Hadas <ahadas@redhat.com>