backend: bugfix in VdsManager on failedToRunVm
Bug description: onVdsDuringFailureTimer assumes that number
of attempts is always >= 0 but this assumption is not always
true. The reason for this is that pausing a quartz job does not
insure it will not run anymore but simply pauses the trigger,
so previously scheduled jobs might still run. This is resolved 
by using simpler one time scheduling.

This patch introduces a behavioral change:
Previously onVdsDuringFailureTimer would reduce failed attempt
by 1 (which is problamatic since it no longer represents failed
attempts when it is decremented by 1 every x time). Previously
there would be a competition between onVdsDuringFailureTimer calls
to failedToRunVm calls. Now whenever attempt limit is reached
number of attempts is set to 0. The effect of this is if we have
a host failing to run vms and those vms were able to run on other hosts
and we detect no oter problem on in (its status is up)
it will keep trying to recover every 30 min (default) regardless.
New behavior is consistent and easy to understand.

Change-Id: Ia4dd140ceecf4954e65ea3f6174a41acea82f6a6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1066693
Signed-off-by: Mooli Tayer <mtayer@redhat.com>