userportal,webadmin: Remove redundant scroll bar in VM general tab
The ProfilesInstanceTypeEditor had the vertical size hardcoded and was wrapped
by a ScrollPanel. This caused the redundant scroll bar in the VM general tab to
appear.

Fixed by removing the ScrollPanel since in all other tabs in the VM dialog the
scrolling is handled by the dialog itself - this change makes the behavior
consistent with the rest of the tabs.

Change-Id: I735fe5fb280f76c6ec771389879433cc9c42ff91
Bug-Url: https://bugzilla.redhat.com/1047883
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>