core: Physically disconnecting blade from chass...
Physically disconnecting blade from chassis does not trigger HA VMs to
restart

This patch verifies that if a Stop PM operation fails and the Host on
which the fencing was performed had running VMs, those VMs will be set
to the UNKNOWN status just after the Stop PM operation failed.

This way, the user can confirm manually that Host has been rebooted and
start HA VMs on another host.

Change-Id: I7d638d574bfe9195e56bdc5970544f97af15af6c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1063432
Signed-off-by: Eli Mesika <emesika@redhat.com>