core: handle 'source' domains when adding a vm from template
When adding a vm from template we need to make sure to choose an
Active domain on which the template disks stored on for the operation-

1. If there's no such domain, a CDA message should be displayed.
2. If the source disk is stored on few domains, the engine will attempt to
select the same domain as the target domain for possibly faster copy.
3. For better user experience, at first we inspect if all the source disks
have a copy on Active domain, later on the destination domains are being
verified and then the source domains will be selected as explained in
(2).

Change-Id: I8d02cfded41a3588dc944d9dfcd5a3eec88c45ab
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1037439
Signed-off-by: Liron Aravot <laravot@redhat.com>