core: Revert Delete jobs that their steps have no async-tasks
Patch 22474 and 26056 added an ability to remove jobs that have
no steps an no async-steps. This caused long running jobs to be
removed while running, causing insert errors and removing them
from customer use.

This patch removes that functionality. If jobs fail, they should
be marked as such. And if they are not, it is the job writer
responsibility to fix this behavior as a bug.

Change-Id: Ief49517c37188839572cd4e95a25b2f921f4b5e1
Bug-Url: https://bugzilla.redhat.com/1076246
Signed-off-by: lzelkha@redhat.com <lzelkha@redhat.com>