core: simplify CommonVmPoolWithVmsCommand#executeCommand
The method CommonVmPoolWithVmsCommand#executeCommand was too long with a
lot of responsibilities inside. This patch extracts the name generation
for the VMs and the creation of the parameters for
AddVmAndAttachToPool to separate methods.

In addition, we are now taking the pool's name from the VmPool object in
the parameters class as we should do, and not from the VM object.

Change-Id: I10b7bd20780f734cbe7c9b7c102b2e5442a722ac
Signed-off-by: Arik Hadas <ahadas@redhat.com>