core: populate groups when getting users by ids
This patch uses group population when getting users by ids. That's needed in order
to get a clear knowledge of the groups (both direct and parent) that a user belongs to.
Also, it makes sure to set proper name and id for these parent groups, in case they don't exist in the DB.

Change-Id: Ief6d9295382fa2b5485e3098589cc20bf9ff61be
Bug-Url: http://bugzilla.redhat.com/1069562
Signed-off-by: Oved Ourfali <oourfali@redhat.com>