core: MacPoolManager strategy, alternative implementation
* introduced new alternative implentation of MacPoolManager
which is less eager browsing its internal structures and
less in need of RAM. There is still need of counting how many
times is each MAC used, but duplicate MAC address is something which
is not expected to occur very often. BitSet is used to store whether
is MAC used or not, and if it has usage higher than 1, pair
MAC<-->count is stored in Map. Counts of user specified MAC outside
from that MAC range are
still stored in HashMap, but modifiable int wrappers are used
instead of java.lang.Integer to cut down calls to HashMap.
*user can input multiple MAC ranges and original implementation
dealt with duplicates using HashSet, which means potentially lot
of unnecessary effort and copying data in RAM from one place to
another. New class RangesWithoutOverlaps was introduced, which
takes user input and transfors it into N intervals
(as few as possible) without overlaps.
*MacPoolManager is singleton; only 'getInstance()' has to be
static other methods should be instance methods. Having them
static is mixing concepts.

Change-Id: I69a5c1b3b43966e49fa6039597c06966ce514618
Bug-Url: https://bugzilla.redhat.com/1063064
Signed-off-by: Martin Mucha <mmucha@redhat.com>