packaging: setup: do not override answer file when applying post install removal
OVESETUP_SYSTEM/nfsConfigEnabled should be removed from post install
file.

we need to find a way to do that in generic way, but for time being we
provide local solution.

the last method we have done to set None at boot stage is good, however
the answer file is configuration just like the post install file, we
have no way to distinguish.

without setting the value at boot stage, the issue would have been
resolved at the 2nd setup run, but we want it to be solved at the next
setup run.

to mark this legacy state a new postinstall variable was added, if it is
None then we know that we have legacy post-install and we do not look
into the content of OVESETUP_SYSTEM/nfsConfigEnabled.

Bug-Url: https://bugzilla.redhat.com/1056777
Change-Id: I907380d1d39c33ffcc43fdd49192a5096a000539
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>