core: Add vdc queries for getting unregistered entities.
A new query which will return if a VM is unregitered or not.
It should be used mainly for REST, since the Import API will be used
also for unregistered VMs,
and there is no indication if the VM is unregistered or simply a VM to
be imported from an Export domain.

Change-Id: I83a937dcc8cd86d7f88f643bb7ba0f03d27a8ee6
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>