webadmin: Replace inner variable references in failure messages
Currently the message processing code ignores variables that were
added by expanding other variables. This is not consistent with
the handling in AuditLog.

This patch fixes that and makes the behaviour the same. The only
risk here is that we have an expansion cycle, but that is
under our control and should not happen.

As an usage example think about the expansion of the following
messages:

"$filterType internal"
"$hostName dev-03"
"$filterName Network"
"$networkNames test"
"$detailMessage network(s) ${networkNames} are missing"
"The host ${hostName} did not satisfy ${filterType} filter
 ${filterName} because ${detailMessage}."

This is related to the our improvement of how we report scheduling
failures.

Change-Id: If70582ae35b8f94b4101dfc33865cc7976fd625c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1077627
Signed-off-by: Martin Sivak <msivak@redhat.com>