engine: Check for conflicting address when hotplugging a disk
When a disk image is hotplugged and it has a stored address we have to
check if that address is not taken by another plugged disk and if that's
the case we should remove the address from the disk that's going to be
hotplugged.

The rationale behind this is: when a disk is hotunplugged qemu may free the
disk's address for further use. If another disk is hotplugged that
address may be given to the new device. Thus would cause that when the
first disk is hotplugged again its address will be rejected because
it's in use by the second disk.

Change-Id: Id5850fd4c15a230dbc27e8c22d47935083da21c2
Bug-Url: https://bugzilla.redhat.com/1079697
Signed-off-by: Xavi Francisco <xfrancis@redhat.com>