core: positive affinity breaks maintenancing a host
I feel that this is best to describe with an example: vm1, vm2 running
with enforcing positive affinity group on host A.
When putting host A to maintenance, both VMs will try to migrate,
and fail because they're violating the affinity constraints.
the host is stuck on preparing for maint.

The fix ignores the host when it's in preparing for maint status.

Change-Id: I7282a9e1c87b30c14869170ed7fb5b9276fad33e
Bug-Url: https://bugzilla.redhat.com/1084794
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
(cherry picked from commit cb27efb5b9561420da8ecb7c44a3b4013c126acb)