vdsbroker: update vmJobs when updating vm statistics
Update vmJob data in the database when updating vmStats.  The jobs
records are created by the initiator of the job (i.e. a command), and
then may be updated and/or removed by VdsUpdateRunTimeInfo.  The job
should be run before the job placeholder is inserted into the database
in order to limit the possibility of a query being made after the
insertion but before job execution, which would result in the job being
prematurely removed from the database.

Change-Id: If1a300d440a0a93381afe047eb294106a38088ec
Bug-Url: https://bugzilla.redhat.com/1058160
Signed-off-by: Greg Padgett <gpadgett@redhat.com>