webadmin: [Fix] Cluster subtabs did not show up
Regression introduced by 3067b72c2b577a6e5d3e061aef38e2e6aebef115. The problem
was that the GlusterStatus parser expected that it will always get a valid
status - but it could get an empty String which caused an exception in
GlusterStatus.valueOf("")

Fixed by retuning null from this parser in case an empty string has to be
parsed. The renderer is able to take care of the null so it is not NPE
vulnerable anymore.

Change-Id: Iad3acc2369f8583b8ba78ed0efb5d1a05755d13c
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>