core:  Distinguish between manual fence resulted..
Distinguish between manual fence resulted from user action and manual
fence called internally by auto fence

Manual fence currently can be executed in 2 flows

1) As part of the non responding treatment when host is restart via its
PM agent.

2) As a result of a manual host stopping and choosing "confirm Host has
been rebooted" from the right-click popup menu

The messages generated by both to the log are of the format:

Manual fence for host ${VdsName} was started.
Manual fence for host ${VdsName} failed.

When this occurs in scenario 1) above, the term "manual" in the audit
log is confusing and should be changed to reflect the fact that was done
internally by the engine code and not by the user

This patch logs different messages to the audit log when resulted from
scenario 1) above.

Change-Id: I274ec95bb3f2c217bd9871afa9d761a1b975d793
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1083750
Signed-off-by: Eli Mesika <emesika@redhat.com>