webadmin: ListModelTypeAheadListBox to not always add to valid values
Added a parameter that determines whether the widget should
automatically add values to the list of "valid" values (those
suggested). This is particularly useful when the selected value is
initialized to some value, that shouldn't be chosen again once it's
been changed (which is the case for the vNIC widget).

Change-Id: Ic35073123696fe7e727464079c31794935a7ad70
Bug-Url: https://bugzilla.redhat.com/1020861
Signed-off-by: Lior Vernia <lvernia@redhat.com>