vdsbroker: read vmJobs update from getVmStats call
For Live Merge, the vmStats return value may optionally contain a list
of vmJob entries.  These jobs indicate new-style async tasks that run on
the VDS.  As long as the job is returned in the stats, it is not
finished.  Once the job is no longer reported, the caller should query
the expected end state to determine if the task was successful.

Change-Id: I6d0ab4372fe98f8d61f4f393e2823d278f33621f
Bug-Url: https://bugzilla.redhat.com/1058160
Signed-off-by: Greg Padgett <gpadgett@redhat.com>