webadmin: migration progress
Adds a new column to the VM main tab which is named "Migration" and
visualzes the progress using the ProgressBarColumn.

Made a special OneColorPercentColumn which is always green (because in contrast
to e.g. memory usage it is not true that the bigger the percentage is, the more
red the color should be).

Change-Id: I31cb4fefe897536b633cce1583881e1e3511c0e8
Bug-Url: https://bugzilla.redhat.com/1083049
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>