restapi: Use current name when updating pool
When updating a VM pool the RESTAPI tries to find the VM data using the
name of the pool, but it uses the name provided by the caller, which may
not be provided, or different to the current one. The effect of this is
that any update not containing the name of the pool fails with a 404
response code. This patch changes that logic to use the current name of
the pool.

Change-Id: Icfd4ae9bf7efed8db6b22cfa7e5dbdbd62cd1136
Related: https://bugzilla.redhat.com/1085386
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>