webadmin: Have minus button on every row in AddRemoveRowWidget
Modified AddRemoveRowWidget to have a minus button on every entry,
including the last one. This to enable removing the last row without
adding a new one first. When removing the last row though, a plus
button needs to be added to the previous one.

Also made sure to add a new ghost entry if the removed entry was the
only one (this is equivalent to reseting the entry's value to a ghost
value).

Some styling changes were required to properly display two buttons per
entry, where this couldn't happen before.

Change-Id: I6f6af3c5a94f7786a9b1e6191f1c36d3181983cf
Bug-Url: https://bugzilla.redhat.com/1065978
Signed-off-by: Lior Vernia <lvernia@redhat.com>