webadmin: Set default vNIC type according to OsInfo
Only set the vNIC type to VirtIO by default if it exists in the OsInfo
vNIC types; if not, pick the first one among the existing values. This
had to be changed for both the vNIC dialog and the VM dialog (where
the vNIC type can't be edited by the user but it is implicitly set
when adding a new vNIC).

Change-Id: I4f09986e661d820b3c10301fa35c840fd9cfe0d8
Bug-Url: https://bugzilla.redhat.com/996044
Signed-off-by: Lior Vernia <lvernia@redhat.com>