userportal,webadmin: call GetVmCustomProperties only once
It did not make sense to call this query in the models since it is a
configuration value.

Solved by using this value the same way as all the other configuration values
are used (e.g. get once in AsyncDataProvider and use it from there).

This patch also refactores all the code to use AsyncDataProvider to get the
value.

Change-Id: I36a472bcf4cd8719967c29626ece4a6afa1fa619
Bug-Url: https://bugzilla.redhat.com/1089938
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>