restapi: Forget HTTP session after invalidate
Currently when the RESTAPI receives a request that contains a valid
session identifier, but that doesn't require persistent authentication
we invalidate the session. This is correct, but we also try to use that
invalidated session later, which generates exceptions. To avoid that
this patch makes sure that when the session is invalidated the variable
that references it is also set to null.

Change-Id: I88dcd37f80eb1fa85aa0441c3cbdfed18e84d5f3
Bug-Url: https://bugzilla.redhat.com/1091371
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>