aaa: Added usage of AuthRecord.VALID_TO
The SessionDataContainer cleans sessions after
UserSessionTimeoutInterval minutes (30 minutes).
AuthRecord.VALID_TO may contain a date in the future
that holds the time the session should expire.
If this time is bigger than the time of the next sessions invalidation
interval,
the sessions invalidation of the session data container
will invalid it anyway.
If it is smaller than the next interval, and a login attempt is performed
using this session, and the expiration time has arrived,
the session will be removed from the session data container, and the user
will be notified the session was timed out.
Change-Id: I53e4a371c1bae8d2480ddd2af921a560c6fe9a85
Topic: AAA
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>