packaging: spec: rework jar/poms installation
there is no reason to install poms that are not part of interface of
product, however I leave this for now.

the timing is due to the fact that we are to expose interface of
extension api and we need to sort out the artifact installation.

the installation of poms was incorrect as the update_maven_depmap at
each subpackage should have been run for subset of files, same for the
pom fragments.

new implementation uses fedora-19 maven integration correctly, while
emulating the minimum required for rhel host, using %files -f syntax.

also add missing jars and poms from build.

Change-Id: I0a0afe8d9c8d85e56fefdb06f155f6459c4ec650
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>