frontend: Always assume GuestAgent to be present for SSO requests
Motivation: There are already so many different names for the guest
agent and their versions that it is an impossible task for the engine
to maintain that. And it's unreliable for such cases. Besides that,
it doesn't tell the engine anyway if there's SSO support or not. Using
the application list to identify the guest agent version is not a good
idea.

Instead we'll check in VDSM whether or not the guest agent is there,
since it gets a heartbeat it'll know early enough if it is responsive
and only then forward the desktopLogin request.
http://gerrit.ovirt.org/27147

Change-Id: Iac7271b4284cc40868ac21ce7eaa4f9ac5e19371
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>