core: SuperUser of DataCenter X cannot approve a host under this Data Center
This patch address the issue above, where you need the following
permissions when approving a host, each from another command that is run in this flow:

1. ChangeVdsCluster (as we change the cluster) - checks for EDIT_HOST_CONFIGURATION
action group both on the host, and the destination cluster
2. UpdateVds (as we can edit other stuff as well) - checks for
EDIT_HOST_CONFIGURATION action group on the host
3. ApproveVdsCommand - checks for CREATE_HOST action group on the host itself,
the cluster, the DC, or System (as I wrote above).

This patch changes the requirement in #1 to include EDIT_HOST_CONFIGURATION on the
host only if the host is NOT in a PendingApproval state, as in the PendingApproval state
there might not be any permissions on the host, as the host is in the Default Cluster, and
the user might not have permissions on it.

Change-Id: Ia59a51e4e3ebf184e0672a077b6dc0d16e60eb8e
Bug-Url: https://bugzilla.redhat.com/1090660
Signed-off-by: Oved Ourfali <oourfali@redhat.com>