[RFE] Improve fencing robustness by retrying...
Improve fencing robustness by retrying failed attempts

This patch adds the following logic to the fencing flow:
In case that a fencing action is failed, an attempt to repeat the failed
action with a different proxy from that used in the failed operation is
done.
In case that no such proxy is available, the operation is repeated with
the same proxy.

Change-Id: I817e8a03695d02ee3f59fbefe0fe41483c3989fe
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=961753
Signed-off-by: Eli Mesika <emesika@redhat.com>