core: Fix UpdateVmDiskCommand resize validation
validateCanResizeDisk() is called from UpdateVmDiskCommand CDA and is
faulty. In it the first two verifications were called whether there was an
attempt to resize the image or not. These two verifications now take
place only if there is indeed an attempt to resize the image.
Need to add a mock in the test.

Change-Id: I760bb3b72d872a1ba5c7af2a9607c7249583652a
Bug-Url: https://bugzilla.redhat.com/1092087
Bug-Url: https://bugzilla.redhat.com/1092371
Signed-off-by: Vered Volansky <vvolansk@redhat.com>