core: Update audit log message when storage is extended.
Update the audit log to indicate that the update of the total size and
free size for the Storage Domain, might take a few moments.
The reason that the operation of extend might suffer from latency issues is that the
mechanism of the storage size is depended on different asynchronous
operations, such as domain monitor in VDSM, IrsBroker in the engine, and
query for storage domains in the GUI main tab.

Change-Id: Id903df200a3325dc1a741d0b543e2cb9466f8a4d
Bug-Url: https://bugzilla.redhat.com/1075515
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>