core: db fix event renaming upgrade script.
db upgrade fails if there's a subscription to VDS_HIGH_NETWORK_USE event
Original upgrade script renamed an event without taking in account that
it may be used by subscribers. This patch updates both event_map and
event_sunscriber tables with the renamed event, in order to do that we
need to remove the FK before the update and adding it back again after
the update is finished.

Change-Id: I3ac7f33fad30aba5cc459ec05dc990a60bbc65e0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1093283
Signed-off-by: Eli Mesika <emesika@redhat.com>