packaging: spec: rework jar/poms installation
at 96199c4e we introduce different method of handling poms/jars, this
derived the need to have real path for %{_javadir}/%{name} as the files
listed in the .mfiles are per that %{name}, so far it used to be symlink
to %{_javadir}/%{engine_name} if package name differ, this is the exact
opposite than what we had before.

however, if previous installation had %{_javadir}%{engine_name} directory
with files that for some reason were not removed, the rpm could not
recreate it as symlink.

as result, we must revert to previous state in this regard, and process
the .mfiles to contain the real path instead of the symlink instead.

Change-Id: Ieb38102400dc861e2369ba8cdbef380593442832
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>