engine: Adjust NetworkUtils to accept any vlan device name- phase 1
The changes in this class will be done in two phases.
Phase 1 (in this patch)-
Adjusting NetworkUtils to get the underling interface name of the nic by
using VdsNetworkInterface.getBaseInterface() instead of determine the
underling nicName from the vlan device name.
In this patch old methods signatures are NOT changed in order to avoid
breaking the compilation. If a change to the signature is needed,
a new method overloads the old one is added.
Phase 2-
Removing the old methods.
Will be done after all the usages of the old methods are
replaced/fixed.

Change-Id: I426159c169b1fa478776abafb38cb7ae6fdac3be
Bug-Url: https://bugzilla.redhat.com/1091863
Bug-Url: https://bugzilla.redhat.com/999975
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>