core: remove jobs of IVdsAsyncCommands on engine startup
Since commands that inherit from IVdsAsyncCommand are not saved to db,
after engine restart we cannot continue to track them. That also means
that the jobs of such commands are not cleaned after engine restarts.

The solution is to remove the jobs of such commands that remained in
running ('STARTED') state. It is safe to do so because the commands are
not restored after the restart so no one will try to access those jobs.

Change-Id: I5973c05f41b90ec48c2961c6f89f5aa110944d6d
Bug-Url: https://bugzilla.redhat.com/1040952
Signed-off-by: Arik Hadas <ahadas@redhat.com>